pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.24.3"
fastapi = "^0.95.2"
python-dotenv = "^1.0.0"
uvicorn = "^0.22.0"
moralis = "^0.1.37"
ethers = "^0.1.1"
web3 = "^6.4.0"
gunicorn = "^20.1.0"
beautifulsoup4 = "^4.12.2"
py-mon = "^1.1.0"
colorama = "^0.4.6"
watchdog = "^3.0.0"
html5lib = "^1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
